+++ src/elements/db/ProductQuery.php
@@ -874,7 +874,7 @@ class ProductQuery extends ElementQuery
     private function _normalizeTypeId()
     {
         if (empty($this->typeId)) {
-            $this->typeId = null;
+            $this->typeId = is_array($this->typeId) ? [] : null;
         } elseif (is_numeric($this->typeId)) {
             $this->typeId = [$this->typeId];
         } elseif (!is_array($this->typeId) || !ArrayHelper::isNumeric($this->typeId)) {